"About" page is done
As the title says,  the"About" page is done. It is still in the same state as main page - not responsive at all, most things are hard-coded. Also there's a problem with centering the main image, I might just go and export a cropped version from Figma to not other with object-fit and object-position.

Also added a working resume, but the current one on google Drive will be replaced a lot of times in the future, because this one was done as a plug for a better one.